split DepositTable file into multiple files
one component per file makes everything more readable!
