+++ src/main/typescript/components/overview/DepositOverview.tsx
@@ -20,7 +20,7 @@ import { AppState } from "../../model/AppState"
 import { DatasetId, Deposit, Deposits } from "../../model/Deposits"
 import { cleanDeposits, deleteDeposit, fetchDeposits } from "../../actions/depositActions"
 import { ReduxAction } from "../../lib/redux"
-import Table from "./DepositTable"
+import DepositTable from "./DepositTable"
 import { Action } from "redux"
 
 interface DepositOverviewProps {
@@ -49,9 +49,9 @@ class DepositOverview extends Component<DepositOverviewProps> {
         const init = loading && <p>loading data...</p>
         const err = loadingError &&
             <p style={{ color: "red" }}>An error occured: {loadingError}. Cannot load data from the server.</p>
-        const data = loaded && <Table deposits={deposits}
-                                      deletingStates={deleting}
-                                      deleteDeposit={this.props.deleteDeposit}/>
+        const data = loaded && <DepositTable deposits={deposits}
+                                             deletingStates={deleting}
+                                             deleteDeposit={this.props.deleteDeposit}/>
 
         return (
             <>
