+++ src/main/typescript/components/overview/DepositTableHead.tsx
@@ -1,0 +1,13 @@
+import * as React from "react"
+
+const DepositTableHead = () => (
+    <tr>
+        <th scope="col">Dataset</th>
+        <th scope="col">Date</th>
+        <th scope="col">State</th>
+        <th scope="col">Notes</th>
+        <th scope="col">Action</th>
+    </tr>
+)
+
+export default DepositTableHead
